requirements.txt
torch~=1.9.0
torchvision~=0.10.0
Pillow~=7.2.0
numpy~=1.19.4
pathlib~=1.0.1
PyYAML~=5.3.1
tqdm~=4.61.2
scikit-learn